fix fastcharge
Due to the missing dtsi definition, this mi,pd-power-max the actual value cannot be read, so it is hard coded. Apollo and alioth both. 33W. At the same time, it fixed the failure to charge quickly due to a long time of measuring unit error. The correct measuring unit is mW, so its actual value should be multiplied by 10. 
Apply commit test, PD certification can be activated normally with original fast charger, charging current 5000ma+, Actual power is 25W+.
Before that, pd fast charge Fail to activate pd could not be activated correctly and the charging current was locked in 1500ma, that is, the actual power is 5W.